refactor: union types can simplify themselves
